backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.3.4
redis==5.0.8
python-dotenv==1.0.1
celery==5.4.0

